Move Query model from website_management into website_analyzer and rename it to SearchKeywords
